Pairing down the decorator code
My goal here is to have the minimum amount of changes needed to get
decorators started while we discuss how the feature should eventually
end up.

This removes the code that handles the decorator interface, the generic
types, and that generates the interface of the decorated class.
Instead, we require that the consumer provides some decorator class with
the correct name and methods and we use that to generate the decorator
class.  Kotlin can infer what the return types are there for the
decorated class.