Added support for errors in callback interfaces (#351)
The Throws attribute is now officially supported on CallbackInterface
methods, allowing callback interfaces to specify which errors they will
throw. If we catch one of those exceptions, we serialize it to the
RustBuffer that's normally used for the return value and send it back to
Rust.

On the Rust side, this means the callback interface trait methods need
to change to return Result<T> instead of T. For now, Throws is optional
and we fall back on the old behavior if it's not present. But this is
deprecated.

We also handle unexpected exceptions in callback interface. If we see
one of those, we stringify it and serialize it to the RustBuffer. This
gets converted into a uniffi::UnexpectedCallbackError in Rust. Users
need to write a From<> impl for that struct to their error enum (unless
they're on the old/deprecated behavior).

Updated callback_interfaces.md to describe the new system, updated the
callbacks example, and implemented it on Python/Kotlin.

Slight change to the FFI return values for callback interface call. See
my comment below for details.  We now recommend returning `1` for all
successful calls.  Before, the documentation suggested returning `0` for
successful calls with void functions, but it doesn't seem like we need a
different code for this since bindings can just not touch the
`RustBuffer`, return `1` and everything works fine.  In fact, some
bindings were doing that.  In particular, this may make the JS
implementation easier.