New checksum handling
Updated the checksum handling to work with the new proc-macro metadata
system.  Now that the metadata is calcualted using const functions
rather than inside the macro code, we don't have a way to include the
checksum in the exported symbol name.

Instead, calculate the metadata and store it as its own symbol.  Then
the bindings code checks the checksums on startup as well as the
contract ABI version.  One nice aspect of this is that we can output a
nicer error message than just "linker error".

- Added a test fixture for this.  Updated the `run_script()` code to support the tests.
- Added flag to allow FFIFunctions to not always have the `&mut
  RustCallStatus` arg.
- Changed the contract version to be a u32 rather than a &str.  This is
  easier to work with as an FFI function.
- Added the `uniffi_meta::ffi_names` module.  This contains the
  canonical code to generate ffi symbol names.
- Use the module path that we calculate in the proc-macro when
  generating the metadata rather than calling `module_path!()`.
  `module_path!()` returns the full module path, but it's more important
  that our metadata agrees with the module path we used whene generating
  the scaffolding.