Better async cancellation
This continues the work from PR#1697 to allow the foreign code to
drop/cancel futures.  Drop causes the future to be immediately dropped,
releasing any resources held -- even if there a Rust executor still holds
a waker.  This is an improvement over the code from PR#1697 which just
made it so they would eventually be dropped.

Cancel is needed for Swift, since withCheckedThrowingContinuation
requires that one of the callbacks is called, even if the task itself is
cancelled.

Separated future construction from starting up the future.  This
simplied the new foreign code -- especially Swift which I couldn't get
working otherwise.  I think there's more than can be simplified, but
I'll wait for a second pass to do that.

It's tricky to implement FFI calls that inputted type-erased
`RustFutureHandle`, since the `F` parameter is an anonymous type that
implements Future. Made a new system that is based to converting an
`Arc<RustFuture<F, T, U>>` into a `Box<Arc<dyn RustFutureFFI>>` before
sending it across the FFI. `Arc<dyn RustFutureFFI>` implements the FFI
and the extra Box converts the wide pointer into a normal pointer.

- Updated the futures fixture tests for this to hold on to the mutex
  longer in the initial call.  This makes it so they will fail unless
  the future is dropped while the mutex is still locked.  Before they
  would only succeed as long as the mutex was dropped once the timeout
  expired.
- Updated the syncronization handling in RustFuture and documented it
  better.
- Updated `UNIFFI_CONTRACT_VERSION` since this is an ABI change
- Updated `RustCallStatus.code` field to be an enum.
- Removed the `RustCallStatus` param from async scaffolding functions.
  These functions can't fail, so there's no need.