Better async cancellation
PR#1697 fixed a memory leak with the futures code which made it so
futures would eventually be dropped correctly.  This makes it so that
when the future is cancelled on the foreign side, we immediately drop
the future -- even if there a Rust executor still has a waker.

This required changing the FFI contract for futures.  Now the foreign
language holds the only permanent strong reference for the future and
frees that reference when the future completes or is cancelled.  The
wakers always store a weak reference.

Separated future construction from starting up the future.  This
simplied the new foreign code a lot -- especially Swift which I couldn't
get working otherwise.  I think there's more than can be simplified, but
I'll wait for a second pass to do that. This also brings us slightly
closer to a potential future discussed in issue#1716 where we can pass
around futures as arguments.

- Updated the futures fixture tests for this to hold on to the mutex
  longer in the initial call.  This makes it so they will fail unless
  the future is dropped while the mutex is still locked.  Before they
  would only succeed as long as the mutex was dropped once the timeout
  expired.
- Updated the bindings code to create 3 FFI functions for each async
  function.
- Created a VTable system so that we can implement the FFI functions
  that input a type-erased RustFuture.
- Updated the syncronization handling in RustFuture and documented it
  better.
- Updated `UNIFFI_CONTRACT_VERSION` since this is an ABI change
- Removed the `RustCallStatus` param from async scaffolding functions.
  These functions can't fail, so there's no need.
- Simplified the `rustfuture.rs` code a bit.  Before, RustFuture needed
  to always be wrapped with pin to ensure that some other module didn't
  accidentally take a &mut to it.  This is no longer needed now that
  it's private.