Refactoring the trait interface code
This is prep work for allowing the foreign bindings to implement traits.

Moved trait interface code in `uniffi_macros` to its own module.

Updated the test trait in the coverall fixture.  I think this one will
work better to test foreign trait impls.  The `ancestor_names()` method
is a good way to test trait implementations that bounce between the Rust
and foreign side of the FFI.

Added Getters test like we have with callback interfaces. We want to
replace callback interfaces with trait interfaces, so we should have
good test coverage.  This actually revealed that the trait code didn't
support exceptions yet.