Foreign-implemented trait interfaces
This adds support for the foreign bindings side to implement Rust traits
and for both sides to pass `Arc<dyn [Trait]>` objects back and forth.

I thought this would be a good time to redo the callback interface code.
The trait interface code uses a foreign-supplied vtable rather than a
single callback and method indexes like the old code used.

Unfortunately, we can't directly return the return value on Python
because of an old ctypes bug (https://bugs.python.org/issue5710).
Instead, input an out param for the return type.  The other main
possibility would be to change `RustBuffer` to be a simple `*mut u8`
(#1779), which would then be returnable by Python.  However, it seems
bad to restrict ourselves from ever returning a struct in the future.
Eventually, we want to stop using `RustBuffer` for all complex data
types and that probably means using a struct instead in some cases.

Renamed `CALL_PANIC` to `CALL_UNEXPECTED_ERROR` in the foreign bindings
templates.  This matches the name in the Rust code and makes more sense
for foreign trait implementations.

-------------------- TODO -----------------------------

This currently requires "wrapping" the object every time it's passed
across the FFI.  If the foreign code receives a trait object, then
passes it back to Rust.  Rust now has a handle to the foreign impl and
that foreign impl just calls back into Rust.  I think #1730 could help
solve this.

I think there should be better tests for reference counts, but I'm
waiting until we address the previous issue to implement them.

Document this