UniFFI bindgen pipeline
This reframes the bindings generation as a compiler pipeline.
I think this is a simpler mental model for people and also allows us to
create some really nice support tools (`peek` and `diff`).

Made the component interface structs generic over a set of types.  This
allows us to transform them into language-specific types as a phase in
the pipeline.  I made it so `Record`, `Enum`, etc. don't get fully
transformed, but instead get an `x` field where languages can store
extra data.  Languages are going to want access to basically all of the
generic fields (name, docstring, fields, variants, etc), so let's not
make them write a bunch of boilerplate to copy the field definitions.